# src/spec/openapi_0.0.1.yaml
openapi: 3.0.0
info:
  title: Control Plane - OpenAPI 3.0
  description: |-
    This is a  Control Plane API Documentation based on the OpenAPI 3.0 specification.
  version: 0.0.1
servers:
  - url: https://controlplane.test.io/api/v1
paths:
...
  /target_policies:
    get:
      tags:
        - TargetPolicies
      summary: Retrieve all available target_policies
      operationId: listtarget_policies
      parameters:
        - required: false
          description: filter Condition for applying filters to the query
          schema:
            type: string
          name: filter
          in: query
      responses:
        "200":
          description: A list of available target_policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TargetPolicyResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
    post:
      tags:
        - TargetPolicies
      summary: Create a new TargetPolicy
      operationId: createTargetPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetPolicyRequest"
      responses:
        "201":
          description: The created TargetPolicy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetPolicyResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "409":
          description: if the TargetPolicy already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "422":
          description: If Request Validation Fails for the Payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
  "/v1/targetpolicies/{app_name}":
    get:
      tags:
      - TargetPolicies
      summary: Get effective Target Policies of application
      description: |-
        Get TargetPolicies Based on app_name and cluster_metadata
        Args:
            app_name  (str): Name of application
            cluster_name  (str, optional): Query to filter TargetPolicies based on cluster name.
            user (User, optional): user (expected role - reader)
        Returns:
            TargetPolicyListResponse: List of TargetPolicies
      operationId: get_effective_target_policies_of_app_v1_targetpolicies__app_name__get
      parameters:
      - required: true
        schema:
          title: App Name
          type: string
        name: app_name
        in: path
      - required: false
        schema:
          title: Cluster Name
          type: string
        name: cluster_name
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TargetPolicyResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  /namespaces:
    get:
      tags:
        - Namespaces
      summary: Retrieve all available namespaces
      operationId: listNamespaces
      responses:
        "200":
          description: A list of available namespaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamespaceListResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
    post:
      tags:
        - Namespaces
      summary: Create a new Namespace
      operationId: createNamespace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NamespaceRequest"
      responses:
        "201":
          description: The created Namespace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "409":
          description: If the Namespace already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "422":
          description: If Request Validation Fails for the Payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
  /namespaces/{namespaceId}:
    get:
      tags:
        - Namespaces
      summary: Retrieve a specific namespace
      operationId: getNamespace
      parameters:
        - name: namespaceId
          in: path
          required: true
          description: The ID of the namespace to retrieve
          schema:
            type: string
      responses:
        "200":
          description: The requested namespace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceResponse"
        "404":
          description: The requested namespace was not found
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "422":
          description: If Request Validation Fails for the Payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
components:
...
  schemas:
    Metadata:
      type: object
...
    TargetPolicyRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the TargetPolicy
        description:
          type: string
          description: A description of the TargetPolicy
        app_selector:
          $ref: "#/components/schemas/Metadata"
        cluster_selector:
          $ref: "#/components/schemas/Metadata"
    TargetPolicyResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the TargetPolicy
        name:
          type: string
          description: The name of the TargetPolicy
        description:
          type: string
          description: A description of the TargetPolicy
        created_by:
          type: string
          description: details of user who created the TargetPolicy
        updated_by:
          type: string
          description: details of user who updated the TargetPolicy
        created_on:
          type: number
          description: created date epoch
        updated_on:
          type: number
          description: updated date epoch
        app_selector:
          $ref: "#/components/schemas/Metadata"
        cluster_selector:
          $ref: "#/components/schemas/Metadata"
    NamespaceRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the Namespace
        description:
          type: string
          description: A description of the Namespace
        cost_center:
          type: string
          description: cost centers
        group:
          type: array
          description: Group names of the team owning this namespace
    NamespaceResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the Namespace
        name:
          type: string
          description: The name of the Namespace
        description:
          type: string
          description: A description of the Namespace
        cost_center:
          type: string
          description: cost centers
        group:
          type: array
          description: Group names of the team owning this namespace
        created_by:
          type: string
          description: details of user who created the Namespace
        updated_by:
          type: string
          description: details of user who updated the Namespace
        created_on:
          type: number
          description: created date epoch
        updated_on:
          type: number
          description: updated date epoch
    NamespaceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NamespaceResponse"
          minItems: 1
          maxItems: 20
          description: The list of Clusters
      additionalProperties: false
      description: A wrapper for the list of Clusters
    ProblemDetails:
      type: object
      properties:
        errorCode:
          description: An application-specific error code, expressed as a string value.
          type: string
        errorDescription:
          description:
            A human-readable explanation specific to this occurrence of
            the problem. Like title, this field's value can be localized.
          type: string
        errorDetails:
          type: array
          items:
            type: object
            properties:
              propertyName:
                description: name of the field in the request which has the error
                type: string
              propertyError:
                description: the error associated  property
                type: string
...
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://controlplane.localhost/oauth/authorize
          tokenUrl: https://controlplane.localhost/oauth/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations